+++ internal/client/facet_query.go
@@ -1,0 +1,56 @@
+package client
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"io"
+	"os"
+)
+
+func (c *Client) FacetQuery(ctx context.Context, params FacetQueryParams) (*FacetQueryResponse, error) {
+	requestParams := map[string]interface{}{
+		"queryType": "facet",
+		"field":     params.Field,
+		"startTime": params.StartTime,
+	}
+
+	if params.Filter != "" {
+		requestParams["filter"] = params.Filter
+	}
+	if params.EndTime != "" {
+		requestParams["endTime"] = params.EndTime
+	}
+	if params.Count > 0 {
+		requestParams["maxCount"] = params.Count
+	}
+	if params.Priority != "" {
+		requestParams["priority"] = params.Priority
+	}
+
+	resp, err := c.makeRequest(ctx, "facetQuery", requestParams)
+	if err != nil {
+		return nil, err
+	}
+	defer resp.Body.Close()
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read response body: %w", err)
+	}
+
+	if c.verbose {
+		fmt.Fprintf(os.Stderr, "Response: %s\n", string(body))
+	}
+
+	var result FacetQueryResponse
+	if err := json.Unmarshal(body, &result); err != nil {
+		return nil, fmt.Errorf("failed to parse response: %w", err)
+	}
+
+	if result.Status != "success" {
+		return nil, fmt.Errorf("API error: %s", result.Message)
+	}
+
+	return &result, nil
+}
